+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @stoker-platform/cli
 
+## 0.5.29
+
+### Patch Changes
+
+- Updated dependencies
+    - @stoker-platform/types@0.5.9
+    - @stoker-platform/node-client@0.5.19
+    - @stoker-platform/utils@0.5.13
+
 ## 0.5.28
 
 ### Patch Changes