+++ packages/node-client/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @stoker-platform/node-client
 
+## 0.5.19
+
+### Patch Changes
+
+- Updated dependencies
+    - @stoker-platform/types@0.5.9
+    - @stoker-platform/utils@0.5.13
+
 ## 0.5.18
 
 ### Patch Changes