+++ packages/web-client/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @stoker-platform/web-client
 
+## 0.5.14
+
+### Patch Changes
+
+- Updated dependencies
+    - @stoker-platform/types@0.5.9
+    - @stoker-platform/utils@0.5.13
+
 ## 0.5.13
 
 ### Patch Changes